move the rewriter to its own files
