+++ src/thorin/transform/cleanup_world.cpp
@@ -2,8 +2,6 @@
 #include "thorin/world.h"
 #include "thorin/analyses/cfg.h"
 #include "thorin/analyses/scope.h"
-#include "thorin/analyses/domtree.h"
-#include "thorin/analyses/scope.h"
 #include "thorin/analyses/verify.h"
 #include "thorin/transform/importer.h"
 #include "thorin/transform/mangle.h"
