+++ src/thorin/transform/mangle.h
@@ -3,26 +3,9 @@
 
 #include "thorin/type.h"
 #include "thorin/analyses/scope.h"
+#include "thorin/transform/rewrite.h"
 
 namespace thorin {
-
-class Rewriter {
-public:
-    explicit Rewriter(World& src, World& dst);
-    explicit Rewriter(World& world) : Rewriter(world, world) {}
-
-    const Def* instantiate(const Def* odef);
-    const Def* insert(const Def* odef, const Def* ndef);
-
-    World& src() { return src_; }
-    World& dst() { return dst_; }
-protected:
-    virtual const Def* rewrite(const Def* odef);
-private:
-    Def2Def old2new_;
-    World& src_;
-    World& dst_;
-};
 
 class Mangler : Rewriter {
 public:
